worker/requirements.txt
runpod>=1.7.0
huggingface_hub>=0.26.0
transformers>=4.46.0
diffusers>=0.32.0
accelerate>=1.1.0
safetensors>=0.4.5
sentencepiece>=0.2.0
einops>=0.8.0
imageio>=2.36.0
imageio-ffmpeg>=0.5.1
av>=13.1.0
pillow>=11.0.0
numpy>=1.26,<2.3
boto3>=1.35.0
requests>=2.32.0
pydantic>=2.9.0
